Drop the confusingly named #table method
It wasn't clear what this method was supposed to do from the name.
Futhermore there is already a method that returns the contents of a
table, `#data`. So I've dropped the `#table` method an added an optional
argument to the `#data` method which is the name of the table you want
to retrieve data from, defaulting to `:data`.